updated restore.sh

+++ copy_from_win.sh
@@ -1,3 +1,13 @@
+#!/bin/bash
+
+if [ -z "$1" ]; then
+  echo "Usage: $0 <windows_path>"
+  exit 1
+fi
+
+WIN_PATH=$1
+LINUX_PATH=$(echo $WIN_PATH | sed 's|\\|/|g' | sed 's|C:|/mnt/c|')
+
 rm -rf backup
 mkdir backup
-cp -r /mnt/c/martes_backup/* ./backup/
+cp -r $LINUX_PATH/* ./backup/
